app move - part2
